requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
apscheduler==3.10.4
click==8.1.7
python-dotenv==1.0.1
Jinja2==3.1.3
python-dateutil==2.8.2
pytest==8.0.1
